researches/MAI-UI/trainer/requirements.txt
# Core dependencies for MAI-UI Training
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Training frameworks
trl>=0.7.0

# Data processing
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0

# Optional: For RL training
# verl  # Uncomment if using verl for RL

# Development
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
